+++ seccardgame/src/migrations/mod.rs
@@ -1,1 +1,2 @@
 pub mod version_one;
+pub mod version_two;